quantity fo test

+++ src/Libraries/Nop.Services/Shipping/GetShippingOptionRequest.cs
@@ -120,7 +120,7 @@ namespace Nop.Services.Shipping
             /// Get quantity
             /// </summary>
             /// <returns></returns>
-            public int GetQuantity()
+            public decimal GetQuantity()
             {
                 if (OverriddenQuantity.HasValue)
                     return OverriddenQuantity.Value;
