+++ src/Presentation/Nop.Web/Areas/Admin/Models/Catalog/ProductModel.cs
@@ -260,7 +260,7 @@ namespace Nop.Web.Areas.Admin.Models.Catalog
         [NopResourceDisplayName("Admin.Catalog.Products.Fields.StockQuantity")]
         public int StockQuantity { get; set; }
 
-        public int LastStockQuantity { get; set; }
+        public decimal LastStockQuantity { get; set; }
 
         [NopResourceDisplayName("Admin.Catalog.Products.Fields.StockQuantity")]
         public string StockQuantityStr { get; set; }
